casi completo
se corrigieron errores de cors, y el front

+++ go/src/modelos/operacion.go
@@ -1,9 +1,9 @@
 package modelos
 
 type Operacion struct {
-	Numero1   int    `json:"numero1"`
-	Numero2   int    `json:"numero2"`
-	Operacion string `json:"operacion"`
-	Resultado int    `json:"resultado"`
-	Fecha     string `json:"fecha"`
+	Numero1   float32 `json:"numero1"`
+	Numero2   float32 `json:"numero2"`
+	Operacion string  `json:"operacion"`
+	Resultado float32 `json:"resultado"`
+	Fecha     string  `json:"fecha"`
 }
